+++ crates/proto/src/version/v975/packets/server_presence_info.rs
@@ -1,8 +1,13 @@
 use bedrockrs_macros::{ProtoCodec, packet};
-use crate::v975::types::PresenceConfiguration;
 
 #[packet(id = 347)]
 #[derive(ProtoCodec, Clone, Debug)]
 pub struct ServerPresenceInfoPacket {
-    pub presence_configuration: PresenceConfiguration,
+    pub presence_configuration: Option<PresenceConfiguration>,
+}
+
+#[derive(ProtoCodec, Clone, Debug)]
+pub struct PresenceConfiguration {
+    pub experience_name: String,
+    pub world_name: String
 }
