+++ crates/proto/src/version/v975/packets/server_store_info.rs
@@ -1,8 +1,13 @@
 use bedrockrs_macros::{ProtoCodec, packet};
-use crate::v975::types::ClientStoreEntryPointConfiguration;
 
 #[packet(id = 346)]
 #[derive(ProtoCodec, Clone, Debug)]
 pub struct ServerStoreInfoPacket {
-    pub client_store_entry_point_configuration: ClientStoreEntryPointConfiguration,
+    pub client_store_entry_point_configuration: Option<ClientStoreEntryPointConfiguration>,
+}
+
+#[derive(ProtoCodec, Clone, Debug)]
+pub struct ClientStoreEntryPointConfiguration {
+    pub store_id: String,
+    pub store_name: String
 }
