--- crates/proto/src/version/v975/types/client_store_entry_point_configuration.rs
@@ -1,7 +1,0 @@
-use bedrockrs_macros::ProtoCodec;
-
-#[derive(ProtoCodec, Clone, Debug)]
-pub struct ClientStoreEntryPointConfiguration {
-    pub store_id: String,
-    pub store_name: String
-}
