+++ crates/proto/src/version/v975/types/mod.rs
@@ -8,7 +8,5 @@ macro_rules! export {
 export!(biome_definition_chunk_gen_data);
 export!(biome_noise_gradient_surface_data);
 export!(biome_surface_builder_data);
-export!(client_store_entry_point_configuration);
 export!(debug_shape);
 export!(dimension_definition_group);
-export!(presence_configuration);
