--- crates/proto/src/version/v975/types/presence_configuration.rs
@@ -1,7 +1,0 @@
-use bedrockrs_macros::ProtoCodec;
-
-#[derive(ProtoCodec, Clone, Debug)]
-pub struct PresenceConfiguration {
-    pub experience_name: String,
-    pub world_name: String
-}
